fix: remove zfs datasets that are not required for initrd
If you use non-legacy mountpoints (which I do), you should not include
`fileSystems` entries in `hardware-configuration.nix` for them except
for the ones that are required for `initrd` and those should have
`options = [ "zfsutil" ];`.

These are the filesystems NixOS needs for boot:
https://github.com/NixOS/nixpkgs/blob/c62f50e86d942d868f7fb96f0450ec0f8ba6bc04/nixos/lib/utils.nix#L10-L14

ZFS will handle mounting datasets automatically if configured to do so,
which NixOS is by default.